Let projections and reactions run as async background processes
In contrast to projections it does not really make sense to have sync reactions as they should be reactions to past events and executed as a new flow. Also if we allow sync reactions we might end up with and endless loop with reactions activating new reactions activation new reactions ect.